moved example config creation into Config to prevent circular import
