+++ lib/Config.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import os from 'os';
 import path from 'path';
 import toml from 'toml';
-import { deepMerge } from './utils/utils';
+import { deepMerge, isPlainObject } from './utils/utils';
 import { PoolConfig } from './p2p/Pool';
 import { LndClientConfig } from './lndclient/LndClient';
 import { RaidenClientConfig } from './raidenclient/RaidenClient';
@@ -149,5 +149,37 @@ class Config {
     return process.env.NODE_ENV === 'production' ? Level.INFO : Level.DEBUG;
   }
 }
+
+export const convertConfigToToml = () => {
+  let result = '';
+
+  const recursivelyConvertJsonToToml = (json: any, prefix: string) => {
+    const nestedPairs: any = [];
+    const simplePairs: any = [];
+
+    if (!json) return;
+
+    Object.keys(json).sort().forEach((key: any) => {
+      if (Object.prototype.toString.call(json[key]) !== '[object Function]') {
+        (isPlainObject(json[key]) ? nestedPairs : simplePairs).push([key, json[key]]);
+      }
+    });
+
+    if (!(prefix === '' || simplePairs.length === 0)) {
+      result += '[' + prefix + ']\n';
+    }
+
+    simplePairs.forEach((value: any) => {
+      result += value[0] + ' = ' + value[1] + '\n';
+    });
+    nestedPairs.forEach((value: any) => {
+      recursivelyConvertJsonToToml(value[1], (prefix === null || prefix === '') ? value[0] : [prefix, value[0]].join('.'));
+    });
+  };
+
+  recursivelyConvertJsonToToml(new Config(), '');
+
+  fs.writeFileSync('../xud.conf', result);
+};
 
 export default Config;
